refactor(core): extract pipeline tests into pipeline/tests.rs
Splits the 2301-line module into 1011 lines of implementation and 1290 of
tests. Behaviour, public API and report schema are unchanged: lib.rs is
byte-identical, and pipeline/mod.rs still exposes exactly compress and
compress_with_estimator.

The test module is declared `mod tests;` -- not `pub mod` -- inside
pipeline/mod.rs. Both details matter. Declaring it in lib.rs instead fails
with 52 name-resolution errors, because the tests reach parent-private
imports (Value, SkippedReason, TransformStatus, WarningCode, TokenEstimator,
ByteHeuristicEstimator) through `use super::*`, and none of those are
re-exported at the crate root. Writing `pub mod` would add
tokenfold_core::pipeline::tests to a published crate's API.

Three test lines are rejoined by rustfmt: dedenting by one level frees four
columns, so expressions that previously had to wrap now fit. Formatting
only.
